♻️ Refactor: MVC 도메인 별 구조 변경
