+++ src/assets/styles/question.scss
@@ -64,6 +64,7 @@
         margin-bottom: 3rem;
         background-image: linear-gradient(45deg, hsl(242, 95%, 50%) 25%, hsl(302, 95%, 50%));
         -webkit-background-clip: text;
+        background-clip: text;
         -webkit-text-fill-color: transparent;
       }
     }
