+++ src/common/IndexedDB.js
@@ -1,6 +1,6 @@
 /* eslint-disable no-console */
 /* eslint-disable no-alert */
-import { randomString } from '../../utils/stringUtils';
+import { randomString } from '../utils/stringUtils';
 import questionJSON from './question.json';
 
 export default class IndexedDB {
