+++ src/manage/ManageViewTemplate.js
